+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- `StateMachine::new` and `StateMachine::new_with_state` are now const functions
 
 ## [v0.6.0] - 2022-11-02